+++ CHANGELOG.md
@@ -15,8 +15,20 @@ All rights reserved.
 ## 0.0.3
 
 - BREAKING CHANGE: stemmer removed from English configuration.
-- BREAKING CHANGE: stemmer incorporated into default tokenFilter for TextAnalyzer.
+- BREAKING CHANGE: stemmer incorporated into default tokenFilter for `TextAnalyzer`.
 
 ## 0.0.4
 
-- Added Tokenizer type definition.
+- Added `Tokenizer` type definition.
+
+## 0.0.5
+
+- BREAKING CHANGE: added `position` property to `Token` class.
+
+## 0.0.6
+
+- Added `TokenCollectionExtension` on `Iterable<Token>`.
+
+## 0.0.7
+
+- Updated documentation.